29Items-Projects/3Python/1E-Commerce-Platform-GLM/backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database & ORM
psycopg[binary]==3.1.16
SQLAlchemy==2.0.25
alembic==1.13.1

# Celery & Redis
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Elasticsearch
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-allauth==0.61.1
PyJWT==2.8.0
cryptography==41.0.7

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
marshmallow==3.20.1

# Data Processing
numpy==1.26.3

# API Documentation
drf-spectacular==0.27.0

# HTTP & API
httpx==0.26.0
requests==2.31.0

# Storage & Media
boto3==1.34.19
Pillow==10.2.0

# Monitoring & Logging
django-prometheus==2.3.1
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

# Code Quality
ruff==0.1.9
black==23.12.1
mypy==1.8.0
django-stubs==4.2.7

# Payment Processing
stripe==7.10.0

# Email
django-anymail==10.2

# Rate Limiting
django-ratelimit==4.1.0

# Django Extensions
django-extensions==3.2.3
